Moved out the static code from the component to utils.
